Add text labels and fix opening a graph
Firstly fixed all graphs to be able to open up a graph
using the right-click context menu. Secondly, I
added the text labels for all injection points. I played
around with the positioning quite a bit, and the text
is added using a for loop in the Plot function. Lastly,
the font size for the title and the reagent texts have
been rescaled after opening a new graph in a new
window or saving it to PNG.