Add images and move computeStats into SA file
Added a folder for holding the figure files and also for holding images
output by these figure files for use in my papers. These images
hold the boxplots for the sensitivity analysis. The computestats function
was removed since it did not need to be defined separately, it was instead
added into the sensitivityAnalysis.m mat file.