Add sensitivity analysis for all sections
Completed sensitivity analysis for not only the whole model,
but for each biological condition as well. This means that
there are 5 sensitivity values for each parameter. Also, to
simplify the structure of the code, the sensitivity analysis
function was placed in the deterministic folder. Results
are now saved to a folder called "SensitivityResults'.